Fix potential dropped mask in AutoregressiveWrapper
`AutoregressiveWrapper.forward` has some logic for replacing a mask. But if `mask` was already the right shape, it gets dropped from `kwargs`.

This PR is an attempt to guess what was intended. I haven't yet convinced myself that the "auto-solve" logic is always what is desired; "explicit is better than implicit" would suggest removing it in favor of an assert with a helpful message (or an exception).

I also removed `pad`, which seemed to be unused (why not use it?).

This is a drive-by PR; I have not tested this code.